==> notebooks/1-preprocessing.ipynb <==
***Imports***

from datasets import load_dataset
import pandas as pd
from nltk.corpus import stopwords

***Loading Dataset***

ds = load_dataset('imdb')

***Creating DataFrames***

train_set = pd.DataFrame(ds['train'])
test_set = pd.DataFrame(ds['test'])

train_set.head()

***Exploring Dataset***

# shape of data
print(train_set.shape)
print(test_set.shape)

train_set.dtypes

train_set['label'].value_counts()

**Insights:** There is `25000` rows of text in both the training data and the test data. There are only 2 features, `label` that labels the data with either `0` for negative and `1` for positive and `text` which is a long string of text. There are both even amounts of data for each label.

***Checking for Nulls***

print(train_set.isnull().sum())
print(test_set.isnull().sum())

**Insights:** There is no null data. So we can continue on with the preprocessing and tokenization

***Removing Punctuation***

def remove_punctuation(text): # Removes punctuation from text
    text = text.lower()
    alphabet = 'abcdefghijklmnopqrstuvwxyz '
    text = "".join([char for char in text if char in alphabet])
    return text

# applying function to data
train_set['text_processed'] = train_set['text'].apply(lambda x: remove_punctuation(x))
test_set['text_processed'] = test_set['text'].apply(lambda x: remove_punctuation(x))

***Tokenization***

def tokenize(text): #  tokenizes the text
    tokens = text.split(" ")
    tokens = ([word for word in tokens if word not in '']) # removing empty spaces
    return tokens

# applying function to data
train_set['text_processed'] = train_set['text_processed'].apply(lambda x: tokenize(x))
test_set['text_processed'] = test_set['text_processed'].apply(lambda x: tokenize(x))

# exploring new tokenized data
print(train_set['text_processed'].apply(lambda x: len(x)).min())
print(train_set['text_processed'].apply(lambda x: len(x)).max())
print(train_set['text_processed'].apply(lambda x: len(x)).mean())

**Insights:** After the tokenizer we can see that the max amount of words is `2460` and minimum is `10`. The mean is `231.7403` showing that there may be outlier data that has significantly more words than the majority of the data. We will be keeping these outliers so the model can be able to read large tex as well.

***Stop Word Removal***

stopwords = stopwords.words('english')

def stop_words(text): # removes common words 
    text = [word for word in text if word not in stopwords]
    return text

# applying function to data
train_set['text_processed'] = train_set['text_processed'].apply(lambda x: stop_words(x))
test_set['text_processed'] = test_set['text_processed'].apply(lambda x: stop_words(x))

# exploring new data after stop word removal
print(train_set['text_processed'].apply(lambda x: len(x)).min())
print(train_set['text_processed'].apply(lambda x: len(x)).max())
print(train_set['text_processed'].apply(lambda x: len(x)).mean())

**Insights:** After removing the stop words the maximum has decreased to `1440` words along with the min decreasing to `4` words. The mean also decreased to `122.9484` words. This is a significant decrease in the range of words.

***Finished Preprocessed Data***

train_set.head()

***Saving Data as CSV's***

train_set.to_csv('../data/train_set.csv', index=False)
test_set.to_csv('../data/test_set.csv', index=False)

Due to github issues I could not save the data into folders like usual. Although the code is here and easy to run to create the data
