==> notebooks/2-representation.ipynb <==
***Imports***

import pandas as pd
from sklearn.feature_extraction.text import CountVectorizer
from sklearn.ensemble import RandomForestClassifier
from sklearn.metrics import f1_score
import joblib

***Load Data***

train_set = pd.read_csv('../data/train_set.csv')
test_set = pd.read_csv('../data/test_set.csv')

***Vectorization***

vectorizer = CountVectorizer()
X_train = vectorizer.fit_transform(train_set['text_processed'])
y_train = train_set['label']
X_test = vectorizer.transform(test_set['text_processed'])
y_test = test_set['label']

***Model***

To start off with this project, we are going to be treating this as a classification problem and using Random Forest to determine the label of each text. This is so we can have a starting model to compare to after we use a pre-trained model.

model = RandomForestClassifier()

model.fit(X_train, y_train)

***Evaluating Model***

The metric for this evaluation will be f1 score. Since it will allow to give a good overall evaluation.

# predicting using test data
y_pred = model.predict(X_test)
results = f1_score(y_test, y_pred)

print(results)

**Insights:** From this f1 score of `0.8492` our model is not doing too bad. Although it can definitely hugely improve. Next notebooks we are gonna be working with pre-trained models.

***Saving the Model***

joblib.dump(model, '../models/random_forest_model.joblib')

Due to github issues none of the models were able to be saved to a folder during this project. But this can be rerun easily if desired.
