==> notebooks/5-optimization.ipynb <==
***Imports***

import pandas as pd
from datasets import load_dataset
from transformers import TrainingArguments, Trainer
from transformers import AutoModelForSequenceClassification
from transformers import AutoTokenizer
from transformers import pipeline
import evaluate 
import numpy as np
from sklearn.metrics import f1_score
from huggingface_hub import notebook_login

***Setting up Login***

notebook_login()

**Note:** For the next model we will be following the same idea as the previous notebook. The only change will be some of the parameters in the training argument.

***Loading Data***

ds = load_dataset('imdb')

***Tokenizer***

tokenizer = AutoTokenizer.from_pretrained('distilbert/distilbert-base-uncased')

# function to tokenize text
def tokenize(batch):
    token_data = tokenizer(batch['text'], truncation=True, padding=True)
    return token_data

# applying functions to data
tokenized_data = ds.map(tokenize, batched=True)

***Model***

model = AutoModelForSequenceClassification.from_pretrained('distilbert/distilbert-base-uncased',
                                                            num_labels=2)

***Evalutation Metric Setup***

# choosing f1 as metric
metric = evaluate.load('f1')

# creating metric functions
def compute_metrics(eval_pred):
    logits, labels = eval_pred
    predictions = np.argmax(logits, axis=-1) # highest prediction
    return metric.compute(predictions=predictions, references=labels)

***Training***

repo_name = 'sentiment-analysis-model'

training_args = TrainingArguments(
    output_dir=repo_name,
    learning_rate=2e-5,
    per_device_train_batch_size=16,
    per_device_eval_batch_size=16,
    num_train_epochs=1,
    weight_decay=0.001, # added weight decay
    warmup_steps=500, # added warmup steps
    push_to_hub=True,
)

trainer = Trainer(
    model=model,
    args=training_args,
    train_dataset=tokenized_data['train'],
    eval_dataset=tokenized_data['test'],
    tokenizer=tokenizer,
    compute_metrics=compute_metrics,
)

trainer.train()

***Evaluating Model***

trainer.evaluate()

**Insights:** We can see from the `f1` score of `0.9264` of this model that is doing pretty well. Although it is ever so slightly worse than our previous model. It is not a significant change. Due to each training taking over an hour to sometimes two hours of my time, I have decided to keep the optimization to only this one. If I had a computer that could run this in a shorter amount of time I would test multiple different optimizations. For the next steps I will be using this model to use on unseen data.

***Pushing Model to Hub***

trainer.push_to_hub()

***Testing Model on Unseen Data***

# getting data and model
text = ds['unsupervised']['text']
my_model = pipeline(model="KiranWood/sentiment-analysis-model")

# trunicating data to fit in the max length for model
max_length = my_model.tokenizer.model_max_length
data = []
for text in text:
    trunicated_text = text[:max_length]
    data.append(trunicated_text)

preds = my_model(data)

# turning into dataframe
preds = pd.DataFrame(preds)

preds.head()

**Insights:** As we can see from this data that it is labeled as either negative or positive. This unseen section of dataset does not have labeled data. So we cannot fully evaluate the final model onto new data. But we can see that it does label new data pretty confidently.
