==> notebooks/6-deployment.ipynb <==
# **Deloyment Steps**

These are the steps that could be taken if this model was to be deployed in a real life setting.

**Fine Tune Model**
- This model would still need a lot of time to be fine tuned and optimized
- Look into prompt tuning for this model

**Evaluate Model**
- Next this model can be tested on different new similar data that is not related to the train/test data
- Identify if there are any issues/bias with the model

**Preparing Deployment**
- Choosing which deployment environment work for the model. Looking into possibly something cloud based.

**DevOps/MLOps**
- Work with DevOps/MLOps if avaiable for a smooth deployment

**Security**
- Implement proper security for the model
- Publish usage guidelines and terms of use

**Deployment**
- Deploy the model 
- Test to make sure that everything is working

**Monitoring**
- Continously monitor to see how well the model is doing after deployment
- Fix any issues that arrise
- Update model to new data as time goes on

**Feedback**
- Gather feedback from users 
- Use the feedback to update and improve the model
- Continue to gather feedback to make improvements 


# **Ethics**

Due to this project being used for learning purposes, it wasn't done with ethics in mind. Although in a real world setting, ethics is very important to keep in mind in all aspects of creating a model. Here are some ethics that could be taken if this project was to be in a real life situation:

- Making sure to source my data ethically
- Avoiding bias such as making sure to get variety of different genres of movies / movies based from different locations across the world
- Insuring the intended use and impact are ethical
- Looking and testing for potential bias during evaluation stage 
- Being transparent about the model and any issues it may have
